btucalccode
You'll notice that i had a little help from chat gpt, as the entire code is wrote in one try in only one page. But to be honest i like programming since i was a child and this was my first project. if you have any recommendation to improve the calc performance or anything lemme know: loldynamo024@gmail.com